% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Exploratory Data Analysis (EDA) - Outlier Treatment}
\usage{
outlier(x)
}
\arguments{
\item{x}{Data Frame that you want to work with}

\item{response}{Dependent(Target) Variable}

\item{seed}{Seed Value, default = 42}

\item{splitRatio}{Ratio to split the dataframe into Train and Test}

\item{minbucket}{Minimum number of observations that should be present in each leaf/terminal node}

\item{cp}{Cost Complexity Paramter Value}

\item{prune}{Pruning Tree on the basis of minimum Cross Validation Error (xerror), default = FALSE}

\item{plot}{Plot the model Tree and CP Table}
}
\value{
A List with Model Attributes & Evaluation Results
}
\description{
This function takes a integer/numerical variable as an input and gives a list with 2 dataframes, one is for the outlier values and one is after removing outliers from that variable.
}
\examples{
create.eval.cart(sleep, group, plot = FALSE)
}
\author{
Gupta, Deepak <deepak@analyticsacedemy.ml>

Sharma, Raman <raman@analyticsacademy.ml>
}
